pyproject.toml
[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | project/bonds/migrations
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.poetry]
name = "origin-test"
version = "0.1.0"
description = ""
authors = ["Steve Rivers <git@rivers.dev>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.1"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.4"
djangorestframework = "^3.11.0"
requests = "^2.22.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-responses = "^0.4.0"
pytest-django = "^3.7.0"
pytest-coverage = "^0.0"
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.761"
ipython = "^7.10.2"
ipdb = "^0.12.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
